requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
diskcache==5.6.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi-slim==0.115.12
h11==0.14.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
kombu==5.5.2
openpyxl==3.1.5
oteapi-core==1.0.0
paramiko==3.5.1
pillow==11.1.0
priority==2.0.0
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
taskgroup==0.2.2
tenacity==9.1.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
